[QoS] Client config (#2690)
* qos config

* respect max backoff itme

* [QoS] [Refactor] Query Weights (#2697)

* query weights

* extra tests

* [QoS] Number of rows per query (#2698)

* num rows

* checkstyle

* fix tests

* no int casting

* fix numRows calculation on batch_mutate

+++ qos-service-impl/src/main/java/com/palantir/atlasdb/qos/QosResource.java
@@ -29,7 +29,7 @@ public class QosResource implements QosService {
     }
 
     @Override
-    public int getLimit(String client) {
-        return config.get().clientLimits().getOrDefault(client, Integer.MAX_VALUE);
+    public long getLimit(String client) {
+        return config.get().clientLimits().getOrDefault(client, Long.MAX_VALUE);
     }
 }
